+++ app/Database/Migrations.hs
@@ -33,6 +33,7 @@ applyMigrations currVersion migrations = do
 migrationList :: [MigrationWrapper]
 migrationList = [ MigrationWrapper {version=2, script=renamePostTables}
                 , MigrationWrapper {version=3, script=renameCoursesTable}
+                , MigrationWrapper {version=4, script=splitTimes}
                 ]
 
 -- | Migration script which renames the Post tables to Program
@@ -46,6 +47,13 @@ renamePostTables = do
 renameCoursesTable :: Migration
 renameCoursesTable =
     addMigration True "ALTER TABLE courses RENAME TO course;"
+
+-- | Migration script to add proper support for year-long courses
+splitTimes :: Migration
+splitTimes = do
+    addMigration True "ALTER TABLE times RENAME COLUMN first_room TO location;"
+    addMigration True "ALTER TABLE times DROP COLUMN second_room;"
+    addMigration True "ALTER TABLE times ADD session varchar(32);"
 
 -- | Gets the current version of the database.
 -- If no version is defined, initialize the