+++ app/Database/Tables.hs
@@ -70,12 +70,12 @@ Meeting
     UniqueMeeting code session section
 
 Times
+    session T.Text Maybe
     weekDay Double
     startHour Double
     endHour Double
     meeting MeetingId
-    firstRoom T.Text Maybe
-    secondRoom T.Text Maybe
+    location T.Text Maybe
 
 Breadth
     description T.Text
@@ -161,20 +161,20 @@ SchemaVersion
 -- ** TODO: Remove these extra types and class instances
 
 data Time' =
-  Time' { weekDay' :: Double,
+  Time' { timeSession' :: Maybe T.Text,
+          weekDay' :: Double,
           startHour' :: Double,
           endHour' :: Double,
-          firstLocation' :: Maybe T.Text,
-          secondLocation' :: Maybe T.Text
+          timeLocation' :: Maybe T.Text
         } deriving (Show, Eq, Generic)
 
 data Time =
-  Time { weekDay :: Double,
-          startHour :: Double,
-          endHour :: Double,
-          firstLocation :: Maybe Building,
-          secondLocation :: Maybe Building
-        } deriving (Show, Generic)
+  Time { timeSession :: Maybe T.Text,
+         weekDay :: Double,
+         startHour :: Double,
+         endHour :: Double,
+         timeLocation :: Maybe Building
+       } deriving (Show, Generic)
 
 -- | A Meeting with its associated Times.
 data MeetTime = MeetTime {meetInfo :: Meeting, timeInfo :: [Time'] }
@@ -229,12 +229,11 @@ instance FromJSON Time' where
 
     building <- o .: "building"
     buildingCode <- building .: "buildingCode"
-    buildingRoomNumber <- building .: "buildingRoomNumber"
-    let meetingRoom1 = Just (T.concat [buildingCode, buildingRoomNumber])
-    meetingRoom2 <- o .:? "assignedRoom2" .!= Nothing
+
+    session <- o .: "sessionCode"
 
     let (adjustedDay, adjustedStartTime, adjustedEndTime) = convertTimeVals meetingDay meetingStartTime meetingEndTime
-    return $ Time' adjustedDay adjustedStartTime adjustedEndTime meetingRoom1 meetingRoom2
+    return $ Time' session adjustedDay adjustedStartTime adjustedEndTime buildingCode
 
 instance FromJSON MeetTime where
   parseJSON (Object o) = do