+++ app/Models/Time.hs
@@ -8,19 +8,18 @@ import Models.Building (getBuilding)
 -- | Convert a Times record into a Time by resolving room codes to Buildings
 buildTime :: Times -> SqlPersistM Time
 buildTime t = do
-    room1 <- getBuilding (timesFirstRoom t)
-    room2 <- getBuilding (timesSecondRoom t)
-    return $ Time (timesWeekDay t)
+    location <- getBuilding (timesLocation t)
+    return $ Time (timesSession t)
+        (timesWeekDay t)
         (timesStartHour t)
         (timesEndHour t)
-        room1
-        room2
+        location
 
 buildTimes :: MeetingId -> Time' -> Times
 buildTimes meetingKey t =
-    Times (weekDay' t)
+    Times (timeSession' t) 
+        (weekDay' t)
         (startHour' t)
         (endHour' t)
         meetingKey
-        (firstLocation' t)
-        (secondLocation' t)
+        (timeLocation' t)